GitHub GraphQL 통신 모듈 분리
